alterando o valor do tempo

+++ app.py
@@ -2,13 +2,14 @@ import RPi.GPIO as GPIO
 import time
 import os
 from dotenv import load_dotenv
+from decimal import Decimal
 
 load_dotenv()
 
 GPIO.setmode(GPIO.BCM)
 
 LED_PIN = int(os.getenv("LED_PIN"))
-TEMPO = int(os.getenv("TEMPO"))
+TEMPO = Decimal(os.getenv("TEMPO"))
 
 GPIO.setup(LED_PIN, GPIO.OUT)
